Three iterations of a Z slicer that returns closed contours.
ZSlicerGrowContour is the most naive one. It grows contours by
linking the segment extremities. It fails as soon as there are
small floating point inacuracies or when an intersection finds
several vertices. The border cases where the plane cuts through
vertices or triangles make it fail.

ZSlicerRoundPositions shows how to deal with border
case in the most brutal way possible: the vertices are shifted
by a certain amount and the Z position of the plane by another,
making sure they never intersect. This seems to be standard in
the publications I have read.

I came to understand that this is a necessary sacrifice if we want
to have semantic, closed contours. We can have another method that
return a soup of segments (not closed) and that would paint them
exactly, but then we won't be able to exploit this for pixel
iteration, not filling the contours when displaying them.

ZSlicerMatchEdges is the best and is close to a relatively
recent publication. It involves building contours by constructing
a graph where the intersecting edges are the vertices. It is the
most robust compared to all other methods.

In most cases the individual contours have the proper orientation.
A hole is painted differently than a contour. But sometimes it is
reversed, which shows when we paint the contours filled. It would
be desirable to have holes properly painted, which means orienting
the polygons with respect to the normals of the triangles that
generate them. I still have to investigate how to do that.