Slice a mesh by a collection of Z planes.
We want to collect the indices of the triangles which minZ is below zr and
maxZ above zr. Because we have to do it for several Zs, it is worth not
iterating through all the triangles every time. The solution here resembles a
a convoluted set intersection.

We sort the triangles by minZ, and perform a binary search to find the
indices of the triangles which minZ is lower than zr (0 -> k1). Then we sort
the resulting indices by maxZ this time, and perform a binary search to find
within these indices the ones that have a maxZ higher than zr (k2 -> k1).

After this we have an indices array, and a lower and higher bound inside array
of triangle that cross the plane zr.