Reslice meshes by a series of XY planes.
Slice a mesh by a Z plane. Only works for meshes that are
two-manifold (otherwise contours are not closed).

Round vertices positions and plane position to even and odd
multiples of asmall fraction of the mesh size, so that they
don't intersect. Build contours by iterating from an intersecting
edge to the adjacent one, with a determined orientation (CCW for
interior contours). The contours returned "know" whether they
limit the interior or exterior of the mesh.